requirements.txt
absl-py==2.2.1
accelerate==1.5.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
black==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
docker-pycreds==0.4.0
executing==2.2.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
gitpython==3.1.44
grpcio==1.71.0
huggingface-hub==0.29.3
hydra-core==1.3.2
idna==3.10
iopath==0.1.9
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.4.2
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.8
markdown==3.7
markupsafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
portalocker==3.1.1
prompt-toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycocotools==2.0.8
pydantic==2.10.6
pydantic-core==2.27.2
pydot==3.0.4
pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentry-sdk==2.24.1
setproctitle==1.3.5
setuptools==78.1.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.1
triton==3.2.0
typing-extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
wandb==0.19.8
wcwidth==0.2.13
werkzeug==3.1.3
wheel==0.45.1
yacs==0.1.8
